Rewrote the incident manager test for mocha.
Tests now pass. All is lintable.
